+++ src/main/java/com/golem/skyblockutils/models/Attribute.java
@@ -3,9 +3,9 @@ package com.golem.skyblockutils.models;
 public class Attribute {
     public final String attribute;
     public final Integer tier;
-    public final Integer price_per;
+    public final Long price_per;
 
-    public Attribute(String attribute, Integer tier, Integer price) {
+    public Attribute(String attribute, Integer tier, Long price) {
         this.attribute = attribute;
         this.tier = tier;
         this.price_per = price >> (tier - 1);
